Giving up on Day 16's problem (I've got other things to before the weekend, and it's getting late).
Brute force attempt for Part 1.

It might work, or maybe not. It solves the example, but takes too long/too much RAM (not sure which comes first) to solve the input I was given (15^15 permutations in the worst case, 11^11 if we remove the valves with flow rates below 10).

Other peoples' solutions seem a lot shorter and smarter. This one completely stumped me.